+++ db/migrate/20170310171656_add_number_to_sites.rb
@@ -1,0 +1,6 @@
+class AddNumberToSites < ActiveRecord::Migration[5.0]
+  def change
+    add_column :sites, :number, :integer
+    add_index :sites, [:number, :project_id], unique: true
+  end
+end